src/main/java/com/jerry/gamemarket/service/Impl/CategoryServiceImpl.java
package com.jerry.gamemarket.service.Impl;
import com.jerry.gamemarket.dao.ProductCategoryDao;
import com.jerry.gamemarket.entity.ProductCategory;
import com.jerry.gamemarket.service.CategoryService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;
import java.util.List;

/*
* 类目service实现类
* author by 李兆杰
* 2018-10-4
* */
@Service
public class CategoryServiceImpl implements CategoryService {
    @Autowired
    private ProductCategoryDao productCategoryDao;
    @Override
    public ProductCategory findOne(Integer categoryId) {
        return productCategoryDao.getOne(categoryId) ;
    }

    @Override
    public List<ProductCategory> findAll() {
        return productCategoryDao.findAll();
    }

    @Override
    public List<ProductCategory> findByCategoryTypeIn(List<Integer> categoryTypeList) {
        return productCategoryDao.findByCategoryTypeIn(categoryTypeList);
    }

    @Override
    public ProductCategory save(ProductCategory productCategory) {
        return productCategoryDao.save(productCategory);
    }
}
